pyproject.toml
[project]
name = "ai-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
